Split mocking between esm-api and esm-emr-api
