+++ packages/framework/esm-api/mock-jest.ts
@@ -1,6 +1,6 @@
 import { of } from 'rxjs';
-import { type SessionStore } from '@openmrs/esm-api';
 import { createGlobalStore } from '@openmrs/esm-state/mock';
+import { type SessionStore } from './src/current-user';
 
 export const setSessionLocation = jest.fn(() => Promise.resolve());
 export const openmrsFetch = jest.fn((url?: string) => new Promise(() => {}));
@@ -13,21 +13,10 @@ export const mockSessionStore = createGlobalStore<SessionStore>('mock-session-st
   session: null,
 });
 export const getSessionStore = jest.fn(() => mockSessionStore);
-export const setCurrentVisit = jest.fn();
-export const newWorkspaceItem = jest.fn();
 export const restBaseUrl = '/ws/rest/v1';
 export const fhirBaseUrl = '/ws/fhir2/R4';
-export const attachmentUrl = '/ws/rest/v1/attachment';
-export const getAttachmentByUuid = jest.fn();
-export const getAttachments = jest.fn();
-export const createAttachment = jest.fn();
-export const deleteAttachmentPermanently = jest.fn();
 export const clearCurrentUser = jest.fn();
 export const refetchCurrentUser = jest.fn();
 export const setUserLanguage = jest.fn();
 export const setUserProperties = jest.fn();
-export const updateVisit = jest.fn();
-export const saveVisit = jest.fn();
-export const getVisitsForPatient = jest.fn();
-export const getStartedVisit = jest.fn();
 export const userHasAccess = jest.fn();
