+++ packages/framework/esm-emr-api/README.md
@@ -1,6 +1,7 @@
 # openmrs-esm-emr-api
 
-openmrs-esm-emr-api exports functions that reflect shared EMR concerns.
+openmrs-esm-emr-api exports functions that reflect shared EMR concerns. Things here are separated out from emr-api
+to remove circular dependencies within the framewok.
 
 See the [Retrieving and posting data](https://o3-docs.openmrs.org/docs/recipes/retrieve-and-post-data)
 page of the Developer Documentation.
