+++ packages/framework/esm-emr-api/mock-jest.ts
@@ -1,0 +1,10 @@
+export const setCurrentVisit = jest.fn();
+export const attachmentUrl = '/ws/rest/v1/attachment';
+export const getAttachmentByUuid = jest.fn();
+export const getAttachments = jest.fn();
+export const createAttachment = jest.fn();
+export const deleteAttachmentPermanently = jest.fn();
+export const updateVisit = jest.fn();
+export const saveVisit = jest.fn();
+export const getVisitsForPatient = jest.fn();
+export const getStartedVisit = jest.fn();
