+++ packages/framework/esm-emr-api/package.json
@@ -14,6 +14,10 @@
     "./src/public": {
       "types": "./src/public.ts",
       "default": "./dist/public.js"
+    },
+    "./mock": {
+      "import": "./mock.ts",
+      "require": "./mock-jest.ts"
     }
   },
   "source": true,
